fix part_1 controler and service

+++ src/resources/users/users.controller.ts
@@ -8,13 +8,14 @@ import {
   Delete,
   UseGuards,
   Req,
+  UnauthorizedException,
 } from '@nestjs/common';
-import { Roles } from './../auth/roles.decorator';
+import { Roles } from '../../auth/roles.decorator';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
-import { JwtGuard } from '../auth/guards/jwt.guard';
-import { RolesGuard } from '../auth/guards/role.guard';
+import { JwtGuard } from '../../auth/guards/jwt.guard';
+import { RolesGuard } from '../../auth/guards/role.guard';
 import { UpdateAdminDto } from './dto/update-admin-dto';
 import { UpdateRoleDto } from './dto/update-role-dto';
 import Role from '../../config/role.enum';
@@ -23,14 +24,7 @@ import Role from '../../config/role.enum';
 export class UsersController {
   constructor(private readonly usersService: UsersService) {}
 
-  //Перенёс в auth до согласования
-  /*
-  @Post('signup')
-  public create(@Body() createUserDto: CreateUserDto): Promise<any> {
-    return this.usersService.create(createUserDto);
-  }*/
-
-  @Roles(Role.SuperAdmin, Role.Admin)
+  @Roles(Role.SUPER_ADMIN, Role.ADMIN)
   @UseGuards(JwtGuard, RolesGuard)
   @Get()
   public findAll(): Promise<any> {
@@ -38,7 +32,7 @@ export class UsersController {
   }
 
   @UseGuards(JwtGuard)
-  @Patch('update')
+  @Patch('me')
   public updateUser(
     @Req() req,
     @Body() updateUserDto: UpdateUserDto,
@@ -49,10 +43,10 @@ export class UsersController {
   @UseGuards(JwtGuard)
   @Get('me')
   public getUser(@Req() req): Promise<any> {
-    return this.usersService.findOneUser(parseInt(req.user.id));
+    return this.usersService.findOne(parseInt(req.user.id));
   }
 
-  @Roles(Role.SuperAdmin)
+  @Roles(Role.SUPER_ADMIN)
   @UseGuards(JwtGuard, RolesGuard)
   @Patch(':id/role')
   public updateSuperAdmin(
@@ -61,32 +55,48 @@ export class UsersController {
   ): Promise<any> {
     return this.usersService.updateRole(parseInt(id), updateUserDto);
   }
-  /* Пока под вопросом. Непонятно будет ли у 
-  администраторов возможность менять email и 
-  сбрасывать пароль у пользователей или нет.
-  */
-  /*
-  @Roles(Role.Admin, Role.SuperAdmin)
+
+  @Roles(Role.SUPER_ADMIN, Role.ADMIN)
+  @UseGuards(JwtGuard, RolesGuard)
+  @Patch(':id/udate-password')
+  public async resetPassword(
+    @Param('id') id: string,
+    @Body() updateUserDto: any,
+  ): Promise<any> {
+    const userBase = await this.findOne(id);
+    if (userBase.role !== Role.SUPER_ADMIN) {
+      return this.usersService.resetPassword(parseInt(id), updateUserDto);
+    } else
+      throw new UnauthorizedException(
+        'Нельзя изменить пароль обратитесь к системному администратору',
+      );
+  }
+
+  @Roles(Role.SUPER_ADMIN, Role.ADMIN)
   @UseGuards(JwtGuard, RolesGuard)
   @Patch(':id')
   public updateAdmin(
     @Param('id') id: string,
-    @Body() updateUserDto: UpdateAdminDto,
+    @Body() updateUserDto: any,
   ): Promise<any> {
     return this.usersService.updateUser(parseInt(id), updateUserDto);
-  }*/
+  }
 
-  @Roles(Role.SuperAdmin, Role.Admin)
+  @Roles(Role.SUPER_ADMIN, Role.ADMIN)
   @UseGuards(JwtGuard, RolesGuard)
   @Get(':id')
   public findOne(@Param('id') id: string): Promise<any> {
     return this.usersService.findOne(parseInt(id));
   }
 
-  @Roles(Role.SuperAdmin, Role.Admin)
+  @Roles(Role.SUPER_ADMIN, Role.ADMIN)
   @UseGuards(JwtGuard, RolesGuard)
   @Delete(':id')
-  public remove(@Param('id') id: string): Promise<any> {
-    return this.usersService.remove(parseInt(id));
+  public async remove(@Param('id') id: string): Promise<any> {
+    const userBase = await this.findOne(id);
+    if (userBase.role !== Role.SUPER_ADMIN) {
+      return this.usersService.remove(parseInt(id));
+    } else
+      throw new UnauthorizedException('Нельзя удалить Супер-Администратора');
   }
 }
